Add initial machinery for MCMC integration

+++ src/math/general.hh
@@ -15,10 +15,10 @@ using Double = double;
 using Degree = Float;
 using Radian = Float;
 
-inline constexpr double PI = 3.14159265358979323846;
-inline constexpr double INV_PI = 0.31830988618379067154;
-inline constexpr double HALF_PI = 1.57079632679489661923;
-inline constexpr double TWO_PI = 6.283185307179586476925;
+inline constexpr Float PI = 3.14159265358979323846; // std::numbers::pi
+inline constexpr Float INV_PI = 0.31830988618379067154;
+inline constexpr Float HALF_PI = 1.57079632679489661923;
+inline constexpr Float TWO_PI = 6.283185307179586476925;
 // inline constexpr double EPSILON = 1e-9; // every func can use its own eps, as it sees fit
 
 template<typename InputType>
